Raise the level cap to 240
Oro's monsters run to level 250 and were out of reach at 180.

Levels 190+ use the 5-term branch of Get_NeedRawEXP, dead code until now. It
multiplied in int and cast to int64 only afterwards, so from 215 the product
wrapped and 16 of the levels between 215 and 240 came out negative -- which the
Add_EXP loop reads as already satisfied. It computes in int64 now. That curve is
kept rather than extended from the gentler cubic because the Oro mobs award
1.5-2.3M exp per kill at 200-220, which only fits requirements in the hundreds
of millions; pacing is tunable through WORLD_VAR_EXP instead.

character.exp becomes bigint, since the requirement passes int32 at level 215.
Same widening for death-penalty exp (server and its client mirror), the mob exp
accumulator, and party exp splitting, whose "* (partyLev + 101)" overflowed
above 17.7M.

+++ database/rose-next.sql
@@ -114,7 +114,7 @@ CREATE TABLE "public"."character" (
     "face_id" integer DEFAULT 1 NOT NULL,
     "hair_id" integer DEFAULT 1 NOT NULL,
     "level" smallint DEFAULT 1 NOT NULL,
-    "exp" integer DEFAULT 0 NOT NULL,
+    "exp" bigint DEFAULT 0 NOT NULL,
     "hp" integer DEFAULT 50 NOT NULL,
     "mp" integer DEFAULT 40 NOT NULL,
     "stamina" integer DEFAULT 5000 NOT NULL,
